# optimized build_attention_bias by 6.37x (637%)