vectorized the nested loops

Resolution (10, 10):
  Original:  0.002841 sec
  Optimized: 0.000555 sec
  Speedup:   5.12x
Resolution (20, 20):
  Original:  0.043811 sec
  Optimized: 0.007248 sec
  Speedup:   6.04x
Resolution (30, 30):
  Original:  0.350406 sec
  Optimized: 0.054987 sec
  Speedup:   6.37x
Resolution (40, 40):
  Original:  0.850569 sec
  Optimized: 0.177181 sec
  Speedup:   4.80x

https://colab.research.google.com/drive/15dW1I_L_gDuwm70ekF7mRdNFQeqPYfhF?usp=sharing
Here's the benchmarking and validation script